css files renamed and imports fixed
