+++ my-app/src/app/components/CourseDesc.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import styles from "./css/CourseDesc.module.css";
+import styles from "@/app/components/css/coursedesc.module.css";
 
 class CourseDesc extends Component {
     render() {
