+++ my-app/src/app/course/page.tsx
@@ -1,11 +1,10 @@
 "use client";
 import React, { Component } from "react";
-import Navbar from "../components/Navbar";
-import CourseDesc from "../components/CourseDesc";
-import Rating from "../components/Rating";
-import Review from "../components/Review";
-import styles from "./course.module.css";
-
+import Navbar from "@/app/components/Navbar";
+import CourseDesc from "@/app/components/CourseDesc";
+import Rating from "@/app/components/Rating";
+import Review from "@/app/components/Review";
+import styles from "@/app/course/course.module.css";
 
 class CoursePage extends Component {
     render() {
